V3 52 0: JSON Log - Star, Half Star, No Star

+++ client/src/store/JSONLogStore.ts
@@ -149,7 +149,7 @@ export default class JSONLogStore {
 	private autoMaxFieldLevel: 1 | 2 = 1;
 	private simpleFields: SimpleFields = { date: '', level: '', category: '', appName: '', kind: '', message: '', rawLine: '' };
 
-	private briefChecked = false;
+	private briefChecked = true;
 	private briefMap: { [key: string]: boolean } = {};
 
 	private rawJsonChecked = false;
@@ -212,7 +212,7 @@ export default class JSONLogStore {
 	@action public toggleBriefChecked() {
 		this.briefChecked = !this.briefChecked;
 		filterStore.filterUpdated();
-		GTag.selectItem('Less Detail Checked', this.briefChecked + '');
+		GTag.selectItem('More Detail Checked', this.briefChecked + '');
 	}
 	public getBriefMap() {
 		return this.briefMap;
@@ -547,10 +547,12 @@ export default class JSONLogStore {
 
 	@action public async deleteEntry(index: number) {
 		const jsonField = this.fields[index];
-		if (jsonField.getName() !== "") {
-			await apFileSystem.deleteFile(jsonField.getDir() + '/' + jsonField.getName());
-		}
 		this.fields.splice(index, 1);
+		if (jsonField.getName() !== "") {
+			if (await apFileSystem.exists(jsonField.getDir() + '/' + jsonField.getName())) {
+				await apFileSystem.deleteFile(jsonField.getDir() + '/' + jsonField.getName());
+			}
+		}
 	}
 }
